some changes

+++ check.py
@@ -1,4 +1,4 @@
 import sys
 
-a, b = int(sys.argv[1]), int(sys.argv[2])
+a, b = float(sys.argv[1]), float(sys.argv[2])
 print(a < b)
